test: move to TypeScript for test file
